file-organizer/backend/requirements.txt
fastmcp @ git+https://github.com/jlowin/fastmcp.git
google-generativeai
python-dotenv
PyMuPDF
Pillow>=10.0.0
rich
typer
fastapi
uvicorn
python-multipart
jinja2
websockets
numpy
watchdog==6.0.0
chromadb
pytesseract
magika
